+++ RELEASENOTES.md
@@ -7,4 +7,4 @@
 - Fixed a crash from the recent upgrade of Realm JS. ([#1034](https://github.com/realm/realm-studio/pull/1034))
 
 ## Internals
-- None
+- Updated the message in the greeting window to include a direct link for signing up via Studios own form. ([#1035](https://github.com/realm/realm-studio/pull/1035))